% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_scores.R
\name{add_simba_scores}
\alias{add_simba_scores}
\title{update an SCE with simba edge scores}
\usage{
add_simba_scores(embs, sce)
}
\arguments{
\item{embs}{output of sce_to_embeddings}

\item{sce}{the associated sce}
}
\value{
SCE updated with altExp giving the simba scores
}
\description{
update an SCE with simba edge scores
}
\examples{
data(nn50)
data(t3k)
t3k = add_simba_scores(nn50, t3k)
t3k
}
