% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_matrix.R
\name{disc_matrix}
\alias{disc_matrix}
\title{discretize a dense matrix}
\usage{
disc_matrix(x, n_bins = 5, q4pos = TRUE)
}
\arguments{
\item{x}{matrix}

\item{n_bins}{numeric(1)}

\item{q4pos}{logical(1) if TRUE, quantiles are formed for positive entries only to form the cutpoints}
}
\value{
matrix with values 0:(nbins-1) corresponding to quantile groups
}
\description{
discretize a dense matrix
}
\note{
The lowest category includes all zero values.
}
