% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_scores.R
\name{embs_to_scores}
\alias{embs_to_scores}
\title{Produce a SummarizedExperiment instance with assay elements
(i,j) the simba score relating gene i to cell j}
\usage{
embs_to_scores(embout, sce, colfeat = "Barcode", rowfeat = "Symbol")
}
\arguments{
\item{embout}{output of `sce_to_embeddings`}

\item{colfeat}{character(1) element of colData of `sce` to index cells}

\item{rowfeat}{character(1) element of rowData of `sce` to index genes}
}
\value{
a SummarizedExperiment
}
\description{
Produce a SummarizedExperiment instance with assay elements
(i,j) the simba score relating gene i to cell j
}
\note{
The simba algorithm produces probability-like scores for
measuring tendency of two entities to be linked.  This
function takes the embeddings produced by sce_to_embeddings
and produces edgewise softmax rescalings of C'G where
C is the embedding for cells and G the embedding for genes
}
