% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_triples.R
\name{make_triples}
\alias{make_triples}
\title{make triples from a discretized SCE}
\usage{
make_triples(sce, fname, colname = "Barcode", assayname = "disc")
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{fname}{character(1) destination file}

\item{colname}{character(1) colData variable to use for labeling cells}

\item{assayname}{character(1) assayNames element to use, defaults to "disc"}
}
\description{
make triples from a discretized SCE
}
\examples{
if (!requireNamespace("TENxPBMCData")) stop("install TENxPBMCData to use this example")
p3k = TENxPBMCData::TENxPBMCData("pbmc3k")
}
