% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trvate.R
\name{make_trvate}
\alias{make_trvate}
\title{convert data.frame to train/test/validate subsamples}
\usage{
make_trvate(dfr, p = c(0.7, 0.2, 0.1))
}
\arguments{
\item{dfr}{a data.frame}

\item{p}{numeric(3) with proportions of samples in train, test, validate}
}
\description{
convert data.frame to train/test/validate subsamples
}
