% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_to_embeddings.R
\name{sce_to_embeddings}
\alias{sce_to_embeddings}
\title{produce torchbiggraph embeddings for discretized single-cell
RNA-seq measures in a SingleCellExperiment}
\usage{
sce_to_embeddings(
  sce,
  workdir,
  N_EPOCHS,
  N_GENES,
  N_GPUS = 1L,
  BATCH_SIZE = 100000L,
  pbg,
  palib
)
}
\arguments{
\item{sce}{SingleCellExperiment instance that includes `disc` among
its assays, intended to hold output of `disc_mat` applied to
normalized log counts}

\item{workdir}{character(1) folder for interim computations, must
exist prior to call}

\item{N_EPOCHS}{numeric(1)}

\item{N_GENES}{numeric(1)}

\item{N_GPUS}{numeric(1)}

\item{BATCH_SIZE}{integer}

\item{pbg}{reference to torchbiggraph}

\item{palib}{reference to pathlib}
}
\value{
a list with elements `cemb` and `gemb` (matrix
representations of embeddings of cells and genes respectively),
`cents` and `gents` (vectors of names of cells and
genes as ordered in the returned matrices),
stats (a data.frame of training statistics),
call (the match.call) and config (a python reference,
not useful after the session in which object was
produced ends).
}
\description{
produce torchbiggraph embeddings for discretized single-cell
RNA-seq measures in a SingleCellExperiment
}
\examples{
pbg = reticulate::import("torchbiggraph")
palib = reticulate::import("pathlib")
data(t3k)
testd = paste0(tempdir(), "/testse")
dir.create(testd)
nn2 = sce_to_embeddings(t3k, testd, N_EPOCHS=2L, BATCH_SIZE=1000L,
   N_GENES=500L, N_GPUS=0L, pbg=pbg, palib=palib)
dir(testd, full.names=TRUE)
}
