% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce2trip.R
\name{sce_to_triples}
\alias{sce_to_triples}
\title{filter a discretized SCE using getTopHVGs from scran, the produce triples (edges)
(cell - weight - gene)}
\usage{
sce_to_triples(sce, outtsv, ngenes = 5000)
}
\arguments{
\item{sce}{instance of SingleCellExperiment with assays logcounts and disc}

\item{outtsv}{character(1) file to hold triples in tsv format}

\item{ngenes}{numeric(1) number of genes to retain from getTopHVGs after
scran::modelGeneVar on logcounts}
}
\description{
filter a discretized SCE using getTopHVGs from scran, the produce triples (edges)
(cell - weight - gene)
}
\examples{
p3k = TENxPBMCData::TENxPBMCData("pbmc3k")
p3k = scater::logNormCounts(p3k)
tsv = tempfile()
mt = sce_to_triples(p3k, tsv, ngenes=2000)
readLines(tsv, 5)
}
