% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_config_schema.R
\name{setup_config_schema}
\alias{setup_config_schema}
\title{make ConfigSchema}
\usage{
setup_config_schema(
  pbgref,
  entities,
  relations,
  entity_path,
  edge_paths,
  checkpoint_path,
  dimension = 400L,
  init_scale = 0.001,
  max_norm = NULL,
  global_emb = FALSE,
  comparator = "dot",
  bias = FALSE,
  loss_fn = "softmax",
  margin = 0.1,
  regularization_coef = 0.001,
  regularizer = "N3",
  init_path = NULL,
  checkpoint_preservation_interval = NULL,
  num_epochs = 50L,
  num_edge_chunks = NULL,
  max_edges_per_chunk = 1000000000L,
  bucket_order = pbgref$config$BucketOrder$INSIDE_OUT,
  workers = NULL,
  batch_size = 1000L,
  num_batch_negs = 50L,
  num_uniform_negs = 1000L,
  disable_lhs_negs = FALSE,
  disable_rhs_negs = FALSE,
  lr = 0.1,
  relation_lr = NULL,
  eval_fraction = 0,
  eval_num_batch_negs = 1000L,
  eval_num_uniform_negs = 1000L,
  background_io = FALSE,
  verbose = 0L,
  hogwild_delay = 2,
  dynamic_relations = TRUE,
  num_machines = 1L,
  num_partition_servers = -1L,
  distributed_init_method = NULL,
  distributed_tree_init_order = TRUE,
  num_gpus = 0L,
  num_groups_for_partition_server = 16L,
  num_groups_per_sharded_partition_server = 1L,
  partition_shard_size = 250L,
  half_precision = FALSE
)
}
\arguments{
\item{pbgref}{instance of torchbiggraph module}

\item{entities}{EntitySchema instance}

\item{relations}{list of RelationSchema instances}
}
\description{
make ConfigSchema
}
\examples{
pbgref = reticulate::import("torchbiggraph")
ent = make_entity_schema(pbgref=pbgref)
entities=reticulate::dict(all=ent)
rs = BiocPBG::make_rel_schema(pbgref=pbgref)
cs = setup_config_schema( pbgref = pbgref, entities = entities,
   relations = list(rs), entity_path = tempdir(), edge_paths = c("tr", "va", "te"),
   checkpoint_path = "cp" )
cs
}
