% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbg.R
\name{tbgR}
\alias{tbgR}
\title{basic interface}
\usage{
tbgR()
}
\value{
basiliskRun result with import from reticulate, typically a Module
}
\description{
basic interface
}
\examples{
tbg <- tbgR()
tbg
names(tbg)
}
