% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbg.R
\name{tbgR}
\alias{tbgR}
\title{basic interface, shows how to get selected references}
\usage{
tbgR()
}
\value{
basiliskRun result with import from reticulate, typically a Module
}
\description{
basic interface, shows how to get selected references
}
\examples{
tbgrefs <- tbgR()
tbgrefs
names(tbgconf$config)
names(tbgconf$train)
}
